+++ src/deprecated/settings/TitleLevelRequests/TitleLevelRequests.js
@@ -2,20 +2,17 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { injectIntl } from 'react-intl';
 
-import { ConfigManager } from '@folio/stripes/smart-components';
 import {
   withStripes,
   TitleManager,
 } from '@folio/stripes/core';
 
+import { ConfigManager } from '../../../settings/components';
 import TitleLevelRequestsForm from './TitleLevelRequestsForm';
 import {
   getInitialValues,
   normalizeData,
 } from './utils';
-import {
-  MODULE_NAMES,
-} from '../../../constants';
 
 export const CONFIG_NAMES = {
   TLR: 'TLR',
@@ -37,7 +34,6 @@ const TitleLevelRequests = ({
     >
       <ConnectedConfigManager
         label={formatMessage({ id: 'ui-circulation.settings.titleLevelRequests.paneTitle' })}
-        moduleName={MODULE_NAMES.SETTINGS}
         configName={CONFIG_NAMES.TLR}
         configFormComponent={TitleLevelRequestsForm}
         stripes={stripes}