+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Fixed new request policy form applied settings from previously edited request policy. Refs UICIRC-1293.
 * Added global permissions for get read-access to values such as tenant’s locale, timezone, and currency. Refs UICIRC-1296.
 * Update css for fix problem visibility Circulation rules. Refs UICIRC-1311.
+* *BREAKING* Migration from mod-config and mod-settings to mod-circulation. Refs UICIRC-1307.
 
 ## [11.0.4](https://github.com/folio-org/ui-circulation/tree/v11.0.4) (2024-12-10)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v11.0.3...v11.0.4)